requirements.txt
fastapi
httpx
pydantic
redis
duckdb
python-dotenv
gpiozero
jinja2
schedule
pyyaml
uvicorn
smtplib
nginx
